Design Changes, added Config Page
I tweaked the design of the footer. I made it cleaner (in both code and
visual) and I added picture icons (Lawyer Toggle on the Left, Config
Page on the right). I also added the load() call to the very end of the
ajax success function in CalendarDb.js, so that it loads as soon as
it's done populating the table for the first time. I also implemented a
function that runs right on doc.ready that checks for the existence of
a table called "appts" and then loads if it finds one. If not, it
allows the sync button to be seen in order to sync the table. I also
added a pagechange function called goTo() because I was tired of
writing $.mobile.pageChange  etc etc. I also added a function that
toggles the Time Frame buttons between 1,3,6,12 months