Extended DB date range
I extended the database range to include all of 2014 so that the
attorneys could test it out in real life (not much use testing only up
to the end of the year). I also changed "Reset" to "ReSync" so that the
attorneys could reload their database, not very efficient, but will
work for now until we implement the unique ID synchronization
techniques. I also took out the alerts for creating the tables and
whatnot, that's more for us than for them.